usb: added usb driver
The USB driver will hold the API for directly interacting with the USB
peripheral. For our purposes, we only use EP0 as TX/RX control, and EP1
as interrupt tx.

The QAZ USB device get successfully recognized on a machine running
Debian (all descriptors are obtained by host), but the transmission of
data has yet to be developed.