usb: finally past enumeration, sending reports
Finally got the device full enumerated and sending reports to the host.
Since the device is defined as a mouse, we are sending cursor move
reports, which successfully occur on the pc.

The problem was not realizing that in the EP_REG macro I was incremented
double the amount I wanted because of pointer arithmetic.

Now, we can fine tune our actual keyboard report descriptor and start
sending it based on key presses.