build: added git hash 'tagged' image copies
When generating the .elf and .bin executables, copies are made that
contain the git hash and dirty status in their name. The repo is
considered 'dirty' if there are uncommitted changes in the index or
unstaged changes in the working tree.

These are useful for keeping track of which version is being flashed
onto the board when using the USB DFU bootloader (done with the user in
mind via STM32CubeProgrammer).