structure: restructuring src files
Source files are now going to be in subdirectories in src/. This
allows modulization of the source files, rather than a big list of
files.

Header guards include this subdir name.

Additionally, the obj and dep files that are creating now mimic this
directory structure (e.g. the object file for src/core.main.c will be
build/obj/core/main.o).