bug: fixed not rebuilding on .hpp change
A typo in Makefile resulted in the dependency files not being included
(was looking in the obj/ dir and not dep/ dir for CXX sources). Thus,
the program would not rebuild on a .hpp change. Oops...

Added clean target to the all target.

Remove stale .bin and .elf files without clean.