persist: added saving user data in flash
A user would likely want their color profile persist over PORs of the
board. In order to do this, config needs to be saved in non-volatile
memory. Since the program is using less than half (in DEBUG) of the
available 32kB of Flash, the last two pages (for a total of 2kB) are
reserved for use with persisting data.

ST provides an EEPROM emulator which handles page
reading/erasing/writing so that reading/writing a single data word is
much easier (like it would be with EEPROM). A small wrapper module
called 'persist' was added to handle interacting with the emulator.

New data words can be added by adding to the PERSIST_DATA_TABLE, and
increasing NB_OF_VAR in eeprom.h

Also removed .h and .c files from cpplint and Doxygen, since they will
likely never conform (especially if from external source).