version: added generated version header
Added a script that takes a template version header and fills in the
variables with build information (git hash, state, board type). The
program can then the definitions for debug information.

This script is ran whenever the project is built.

The debug character array is used to write the hash/state of the repo
into persistent memory, so the flash can be peeked into to determine
information about the program.