bugfix: overindexing array
When the maximum number of keys have been read, only the row iterating
loop is broken out of, when both the column and row loops needed to
exited. If 6 keys were found active, but another was active on the next
column, the key buffer would be accessed as if it was able to hold 7
keys.