is31fl3746a: new RGB LED driver
The v1.1 QAZ 65% will be using a new RGB LED driver, which drives
LEDs in a matrix style. This device is controlled over I2C.

Instead of controlling entire banks of LEDs, each individual LED can be
controlled.

In comparison to the LP5009, the device does not automatically control
LED intensity when brightness or color is changed individually, so
the must be handled in software.