i2c_speed: icreased to fast most, 400kHz
Since so much time is spent writing to the I2C LED driver, the task can
actually overrun. Thus, the speed was increased to 400kHz. Additionally,
to make the slow color/brightness changes more smooth, decrease the task
period for the lighting (we need to decrease the timeslice period as
well).