bsp: splitting of program based on bsp
Well, this isn't a very good solution but it will work for now.

I want to be able to easily switch in and out modules based on
whether a particular board needs it. For example, the QAZ 65%
needs lighting, the key matrix, and keyboard hid functionality. But the
QAZ Media board doesn't need these.

So how do I handle this? In a very non-scalable way. Each BSP has its
own init which calls the appropriate inits, great. And the non-relevant
content in the bsp header can be removed, great. But we are still
compiling sources we don't need which causes compiler errors because we
removed that content from the header. So now we also need to switch what
we compile.

But again, this works for now, while I develop the QAZ Media board, and
think of a better way.