mute: add mute led toggling
When the volume knob is pressed, the mute LED will toggle.

Unfortunately, this isn't really the best use of the LED. The MUTE
LED should reflect the MUTE state of the machine (which can be changed
on the computer, as well). However, the state of this LED, as opposed to
the capslock/numlock/etc LEDs, are not reported by standard
HID drivers.