user config: add local user config makefile
Added a template for a user config makefile that is not actually tracked
by Git. The user needs to create this file (ideally by copying the
template) so that these variables can be used when building.

The reason for making this something not tracked is because we really
don't want to track these sort of changes, and force stashing/removing
these changes when committing, since there isn't a "default" build type.