qaz 65% v1.2: in-progress, updated pcb
An issue seen with v1.1 (and v1.0 to a degree) was board warping. It's
suspected that a copper imbalance causes this, since there is a full
copper pour on the back of the board, and comparatively little copper on
the front of the board.

A ground pour was added to the top. This is desired over floating
copper, but requires many ground via to stitch the pour to ground, and
to prevent any antennae-like patches.

Also slightly moved some of the in-board screw holes so that they don't
brush up so close to key switches.